move theme to current_user
