+++ openbb_terminal/common/behavioural_analysis/reddit_view.py
@@ -15,8 +15,7 @@ import seaborn as sns
 
 from openbb_terminal import OpenBBFigure
 from openbb_terminal.common.behavioural_analysis import reddit_model
-from openbb_terminal.config_terminal import theme
-from openbb_terminal.core.session.current_user import get_current_user
+from openbb_terminal.core.session.current_user import get_current_user, theme
 from openbb_terminal.decorators import check_api_key, log_start_end
 from openbb_terminal.helper_funcs import export_data, plot_autoscale, print_rich_table
 from openbb_terminal.rich_config import console
