+++ openbb_terminal/common/behavioural_analysis/twitter_view.py
@@ -11,7 +11,7 @@ from dateutil import parser as dparse
 
 from openbb_terminal import OpenBBFigure
 from openbb_terminal.common.behavioural_analysis import twitter_model
-from openbb_terminal.config_terminal import theme
+from openbb_terminal.core.session.current_user import theme
 from openbb_terminal.decorators import log_start_end
 from openbb_terminal.helper_funcs import export_data, get_closing_price
 from openbb_terminal.rich_config import console
