+++ openbb_terminal/config_terminal.py
@@ -10,15 +10,6 @@ from openbb_terminal.core.config.paths import MISCELLANEOUS_DIRECTORY
 from openbb_terminal.core.plots.backend import plots_backend
 from openbb_terminal.core.session.current_user import get_current_user
 
-from .helper_classes import TerminalStyle as _TerminalStyle
-
-# # Terminal UX section
-current_user = get_current_user()
-theme = _TerminalStyle(
-    current_user.preferences.MPL_STYLE,
-    current_user.preferences.PMF_STYLE,
-    current_user.preferences.RICH_STYLE,
-)
 
 # Start Backend for plotting
 plots_backend().start(load_env_vars("DEBUG_MODE", strtobool, False))
