+++ openbb_terminal/cryptocurrency/defi/llama_view.py
@@ -7,9 +7,9 @@ from typing import Optional, Union
 
 from openbb_terminal import (
     OpenBBFigure,
-    config_terminal as cfg,
 )
 from openbb_terminal.core.plots.backend import plots_backend
+from openbb_terminal.core.session.current_user import theme
 from openbb_terminal.cryptocurrency.cryptocurrency_helpers import read_data_file
 from openbb_terminal.cryptocurrency.defi import llama_model
 from openbb_terminal.decorators import log_start_end
@@ -47,7 +47,7 @@ def display_grouped_defi_protocols(
     )
     fig.set_title(f"Top {limit} dApp TVL grouped by chain")
 
-    colors = iter(cfg.theme.get_colors(reverse=True))
+    colors = iter(theme.get_colors(reverse=True))
 
     for chain in chains:
         chain_filter = df.loc[df.Chain == chain]
