+++ openbb_terminal/cryptocurrency/nft/nftpricefloor_view.py
@@ -7,7 +7,7 @@ import logging
 import os
 from typing import Optional, Union
 
-from openbb_terminal.config_terminal import theme
+from openbb_terminal.core.session.current_user import theme
 from openbb_terminal import OpenBBFigure
 from openbb_terminal.cryptocurrency.nft import nftpricefloor_model
 from openbb_terminal.decorators import log_start_end
