+++ openbb_terminal/cryptocurrency/onchain/shroom_view.py
@@ -6,7 +6,7 @@ import os
 from typing import Optional, Union
 
 from openbb_terminal import OpenBBFigure
-from openbb_terminal.config_terminal import theme
+from openbb_terminal.core.session.current_user import theme
 from openbb_terminal.cryptocurrency.onchain.shroom_model import (
     get_daily_transactions,
     get_dapp_stats,
