+++ openbb_terminal/cryptocurrency/overview/blockchaincenter_view.py
@@ -5,7 +5,7 @@ from datetime import datetime
 from typing import Optional, Union
 
 from openbb_terminal import OpenBBFigure
-from openbb_terminal.config_terminal import theme
+from openbb_terminal.core.session.current_user import theme
 from openbb_terminal.cryptocurrency.overview.blockchaincenter_model import (
     DAYS,
     get_altcoin_index,
