+++ openbb_terminal/cryptocurrency/overview/glassnode_view.py
@@ -7,7 +7,7 @@ import numpy as np
 import pandas as pd
 
 from openbb_terminal import OpenBBFigure
-from openbb_terminal.config_terminal import theme
+from openbb_terminal.core.session.current_user import theme
 from openbb_terminal.cryptocurrency.overview.glassnode_model import get_btc_rainbow
 from openbb_terminal.decorators import check_api_key, log_start_end
 from openbb_terminal.helper_funcs import export_data
