+++ openbb_terminal/cryptocurrency/overview/loanscan_view.py
@@ -7,8 +7,8 @@ import pandas as pd
 
 from openbb_terminal import (
     OpenBBFigure,
-    config_terminal as cfg,
 )
+from openbb_terminal.core.session.current_user import theme
 from openbb_terminal.cryptocurrency.overview import loanscan_model
 from openbb_terminal.decorators import log_start_end
 from openbb_terminal.helper_funcs import export_data, print_rich_table
@@ -63,7 +63,7 @@ def display_crypto_rates(
     df_non_null = pd.melt(df.reset_index(), id_vars=["index"]).dropna()
 
     assets = df_non_null.variable.unique().tolist()
-    colors = iter(cfg.theme.get_colors(reverse=True))
+    colors = iter(theme.get_colors(reverse=True))
 
     for asset in assets:
         width = df_non_null.loc[(df_non_null.variable == asset)]
