+++ openbb_terminal/economy/alphavantage_view.py
@@ -6,7 +6,7 @@ import os
 from typing import Optional, Union
 
 from openbb_terminal import OpenBBFigure
-from openbb_terminal.config_terminal import theme
+from openbb_terminal.core.session.current_user import theme
 from openbb_terminal.decorators import check_api_key, log_start_end
 from openbb_terminal.economy import alphavantage_model
 from openbb_terminal.helper_funcs import export_data, print_rich_table
