+++ openbb_terminal/parent_classes.py
@@ -21,11 +21,11 @@ from prompt_toolkit.formatted_text import HTML
 from prompt_toolkit.styles import Style
 from rich.markdown import Markdown
 
-# IMPORTS INTERNAL
-from openbb_terminal.config_terminal import theme
 from openbb_terminal.core.completer.choices import build_controller_choice_map
 from openbb_terminal.core.session.constants import REGISTER_URL
-from openbb_terminal.core.session.current_user import get_current_user, is_local
+
+# IMPORTS INTERNAL
+from openbb_terminal.core.session.current_user import get_current_user, is_local, theme
 from openbb_terminal.cryptocurrency import cryptocurrency_helpers
 from openbb_terminal.custom_prompt_toolkit import NestedCompleter
 from openbb_terminal.decorators import log_start_end
