+++ openbb_terminal/portfolio/portfolio_optimization/optimizer_view.py
@@ -24,7 +24,7 @@ from matplotlib.lines import Line2D
 
 # IMPORTS INTERNAL
 from openbb_terminal.core.session.current_user import get_current_user
-from openbb_terminal.config_terminal import theme
+from openbb_terminal.core.session.current_user import theme
 from openbb_terminal.decorators import log_start_end
 from openbb_terminal.helper_funcs import plot_autoscale, print_rich_table
 from openbb_terminal.portfolio.portfolio_optimization import (
