+++ openbb_terminal/rich_config.py
@@ -16,7 +16,7 @@ from openbb_terminal import (
     config_terminal as cfg,
 )
 from openbb_terminal.core.config.paths import MISCELLANEOUS_DIRECTORY
-from openbb_terminal.core.session.current_user import get_current_user
+from openbb_terminal.core.session.current_user import get_current_user, theme
 
 # pylint: disable=no-member,c-extension-no-member
 
@@ -24,7 +24,7 @@ from openbb_terminal.core.session.current_user import get_current_user
 # https://rich.readthedocs.io/en/stable/appendix/colors.html#appendix-colors
 # https://rich.readthedocs.io/en/latest/highlighting.html#custom-highlighters
 
-CUSTOM_THEME = Theme(cfg.theme.console_style)
+CUSTOM_THEME = Theme(theme.console_style)
 
 RICH_TAGS = [
     "[menu]",
