+++ openbb_terminal/stocks/behavioural_analysis/finnhub_view.py
@@ -9,7 +9,7 @@ import pandas as pd
 import yfinance as yf
 
 from openbb_terminal import OpenBBFigure
-from openbb_terminal.config_terminal import theme
+from openbb_terminal.core.session.current_user import theme
 from openbb_terminal.decorators import check_api_key, log_start_end
 from openbb_terminal.helper_funcs import export_data
 from openbb_terminal.stocks.behavioural_analysis import finnhub_model
