+++ openbb_terminal/stocks/comparison_analysis/finbrain_view.py
@@ -9,7 +9,7 @@ import numpy as np
 import pandas as pd
 
 from openbb_terminal import OpenBBFigure
-from openbb_terminal.config_terminal import theme
+from openbb_terminal.core.session.current_user import theme
 from openbb_terminal.decorators import log_start_end
 from openbb_terminal.helper_funcs import export_data, print_rich_table
 from openbb_terminal.rich_config import console
