+++ openbb_terminal/stocks/fundamental_analysis/yahoo_finance_view.py
@@ -10,7 +10,7 @@ import pandas as pd
 import yfinance as yf
 
 from openbb_terminal import OpenBBFigure
-from openbb_terminal.config_terminal import theme
+from openbb_terminal.core.session.current_user import theme
 from openbb_terminal.decorators import log_start_end
 from openbb_terminal.helper_funcs import (
     export_data,
