+++ openbb_terminal/stocks/insider/businessinsider_view.py
@@ -10,7 +10,7 @@ from typing import Optional
 import pandas as pd
 
 from openbb_terminal import OpenBBFigure
-from openbb_terminal.config_terminal import theme
+from openbb_terminal.core.session.current_user import theme
 from openbb_terminal.decorators import log_start_end
 from openbb_terminal.helper_funcs import (
     export_data,
