+++ openbb_terminal/stocks/options/screen/syncretism_view.py
@@ -7,7 +7,7 @@ import os
 from typing import List, Optional, Union
 
 from openbb_terminal import OpenBBFigure
-from openbb_terminal.config_terminal import theme
+from openbb_terminal.core.session.current_user import theme
 from openbb_terminal.decorators import log_start_end
 from openbb_terminal.helper_funcs import export_data, print_rich_table
 from openbb_terminal.rich_config import console
